Rename domain pkg to shared
